Modernize workspace build and API stack
