+++ README.md
@@ -23,11 +23,11 @@ https://www.youtube.com/watch?v=cCBdGGHIX-0&t=2s
 
 A more detailed guide will be released soon but for now, follow these instructions:
 
-1. Install NodeJS v18 and Yarn (with npm `npm install -g yarn`)
+1. Install NodeJS v22 and enable pnpm with `corepack enable`.
 2. Clone this repo
 3. Open two Terminals/CMD in the directory with extracted source
-4. Run this command on one window to start the server: `cd server && yarn install && yarn start`
-5. Run this command on the second terminal to start the client: `cd client && yarn install && yarn start`
+4. Install every workspace once from the repository root: `pnpm install`.
+5. Run `pnpm dev:server` in one terminal and `pnpm dev:client` in another.
 6. Go to `localhost:8000` (server hosted at `localhost:3000`)
 
 If you face any issues, you can get help on the [Swordbattle.io Discord Server](https://discord.com/invite/BDG8AfkysZ)!
