+++ client/package.json
@@ -2,177 +2,63 @@
   "name": "swordbattle.io-client",
   "version": "0.1.0",
   "private": true,
-  "homepage": "./",
+  "type": "module",
   "scripts": {
-    "prestart": "node ../scripts/updatebanners.js",
-    "start": "node scripts/start.js",
-    "analyze": "source-map-explorer 'build/static/js/*.js'",
+    "predev": "node ../scripts/updatebanners.js",
+    "dev": "vite --host 0.0.0.0",
+    "start": "pnpm dev",
     "prebuild": "node ../scripts/updatebanners.js",
-    "build": "node scripts/build.js",
-    "build-protocol": "npx pbjs --ts ./src/game/network/Protocol.ts ./src/game/network/schema.proto",
-    "test": "node scripts/test.js"
+    "build": "tsc --noEmit && vite build",
+    "preview": "vite preview --host 0.0.0.0",
+    "build-protocol": "pnpm exec pbjs --ts ./src/game/network/Protocol.ts ../packages/shared/protocol/schema.proto",
+    "typecheck": "tsc --noEmit",
+    "test": "vitest run --passWithNoTests"
   },
   "dependencies": {
-    "@babel/core": "^7.16.0",
     "@fortawesome/fontawesome-svg-core": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
     "@reduxjs/toolkit": "^1.9.7",
-    "@svgr/webpack": "^5.5.0",
-    "babel-jest": "^27.4.2",
-    "babel-loader": "^8.2.3",
-    "babel-plugin-named-asset-import": "^0.3.8",
-    "babel-preset-react-app": "^10.0.1",
-    "bfj": "^7.0.2",
+    "@swordbattle/shared": "workspace:*",
     "bootstrap": "^5.3.2",
-    "browserslist": "^4.18.1",
-    "camelcase": "^6.2.1",
-    "case-sensitive-paths-webpack-plugin": "^2.4.0",
     "chart.js": "^4.4.1",
     "clsx": "^2.0.0",
-    "css-loader": "^6.5.1",
-    "css-minimizer-webpack-plugin": "^3.2.0",
     "date-fns": "^3.3.1",
     "dompurify": "3.2.4",
-    "dotenv": "^10.0.0",
-    "dotenv-expand": "^5.1.0",
-    "eslint": "^8.3.0",
-    "eslint-config-react-app": "^7.0.1",
-    "eslint-webpack-plugin": "^3.1.1",
-    "file-loader": "^6.2.0",
-    "fs-extra": "^10.0.0",
-    "html-webpack-plugin": "^5.5.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.4.3",
-    "jest-resolve": "^27.4.2",
-    "jest-watch-typeahead": "^1.0.0",
     "marked": "12.0.0",
-    "mini-css-extract-plugin": "^2.4.5",
+    "earcut": "^2.2.4",
     "phaser": "^3.60.0",
     "phaser3-rex-plugins": "^1.60.2",
     "pixi.js": "7",
     "pixi.js-legacy": "^7.4.3",
-    "postcss": "^8.4.4",
-    "postcss-flexbugs-fixes": "^5.0.2",
-    "postcss-loader": "^6.2.1",
-    "postcss-normalize": "^10.0.1",
-    "postcss-preset-env": "^7.0.1",
-    "prompts": "^2.4.2",
     "react": "^18.2.0",
-    "react-app-polyfill": "^3.0.0",
     "react-chartjs-2": "^5.2.0",
     "react-countup": "^6.5.0",
-    "react-dev-utils": "^12.0.1",
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.3",
-    "react-refresh": "^0.11.0",
     "react-router-dom": "^6.17.0",
-    "resolve": "^1.20.0",
-    "resolve-url-loader": "^4.0.0",
     "sass": "^1.63.3",
-    "sass-loader": "^12.3.0",
-    "semver": "^7.3.5",
-    "source-map-explorer": "^2.5.3",
-    "source-map-loader": "^3.0.0",
-    "style-loader": "^3.3.1",
-    "tailwindcss": "^3.0.2",
-    "terser-webpack-plugin": "^5.2.5",
-    "typescript": "^5.1.3",
-    "valtio": "^1.10.5",
-    "webpack": "^5.64.4",
-    "webpack-dev-server": "^4.6.0",
-    "webpack-manifest-plugin": "^4.0.2",
-    "workbox-webpack-plugin": "^6.4.1"
+    "valtio": "^1.10.5"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@types/jest": "^29.5.2",
     "@types/node": "^20.3.1",
+    "@types/earcut": "^2.1.4",
     "@types/react": "^18.2.12",
     "@types/react-dom": "^18.2.5",
-    "@typescript-eslint/eslint-plugin": "^5.59.11",
-    "@typescript-eslint/parser": "^5.59.11",
-    "eslint": "^8.42.0",
-    "protobufjs": "^7.2.5"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "@vitejs/plugin-react": "^5.0.4",
+    "autoprefixer": "^10.4.21",
+    "jsdom": "^26.1.0",
+    "postcss": "^8.5.6",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-preset-env": "^10.3.1",
+    "protobufjs": "^7.2.5",
+    "protobufjs-cli": "^1.1.3",
+    "typescript": "^5.9.2",
+    "vite": "^7.1.7",
+    "vite-plugin-svgr": "^4.5.0",
+    "vitest": "^3.2.4"
   },
   "engines": {
-    "node": ">=18.x"
-  },
-  "jest": {
-    "roots": [
-      "<rootDir>/src"
-    ],
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx,ts,tsx}",
-      "!src/**/*.d.ts"
-    ],
-    "setupFiles": [
-      "react-app-polyfill/jsdom"
-    ],
-    "setupFilesAfterEnv": [],
-    "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
-      "<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}"
-    ],
-    "testEnvironment": "jsdom",
-    "transform": {
-      "^.+\\.(js|jsx|mjs|cjs|ts|tsx)$": "<rootDir>/config/jest/babelTransform.js",
-      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^(?!.*\\.(js|jsx|mjs|cjs|ts|tsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs|cjs|ts|tsx)$",
-      "^.+\\.module\\.(css|sass|scss)$"
-    ],
-    "modulePaths": [],
-    "moduleNameMapper": {
-      "^react-native$": "react-native-web",
-      "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
-    },
-    "moduleFileExtensions": [
-      "web.js",
-      "js",
-      "web.ts",
-      "ts",
-      "web.tsx",
-      "tsx",
-      "json",
-      "web.jsx",
-      "jsx",
-      "node"
-    ],
-    "watchPlugins": [
-      "jest-watch-typeahead/filename",
-      "jest-watch-typeahead/testname"
-    ],
-    "resetMocks": true
-  },
-  "babel": {
-    "presets": [
-      "react-app"
-    ]
-  },
-  "volta": {
-    "node": "22.23.1",
-    "yarn": "1.22.22"
+    "node": ">=22"
   }
 }
