+++ client/postcss.config.cjs
@@ -1,0 +1,9 @@
+module.exports = {
+  plugins: {
+    'postcss-flexbugs-fixes': {},
+    'postcss-preset-env': {
+      autoprefixer: { flexbox: 'no-2009' },
+      stage: 3,
+    },
+  },
+};
