+++ client/src/react-app-env.d.ts
@@ -1,3 +1,5 @@
+/// <reference types="vite/client" />
+/// <reference types="vite-plugin-svgr/client" />
 /// <reference types="node" />
 /// <reference types="react" />
 /// <reference types="react-dom" />
@@ -35,17 +37,6 @@ declare module '*.jpeg' {
 }
 
 declare module '*.png' {
-  const src: string;
-  export default src;
-}
-
-declare module '*.svg' {
-  import * as React from 'react';
-
-  export const ReactComponent: React.FunctionComponent<React.SVGProps<
-    SVGSVGElement
-  > & { title?: string }>;
-
   const src: string;
   export default src;
 }
