+++ packages/shared/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@swordbattle/shared",
+  "version": "1.0.0",
+  "private": true,
+  "main": "index.js",
+  "types": "index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./index.d.ts",
+      "default": "./index.js"
+    },
+    "./protocol": {
+      "types": "./protocol/index.d.ts",
+      "default": "./protocol/index.js"
+    },
+    "./protocol/schema.proto": "./protocol/schema.proto"
+  },
+  "files": ["index.js", "index.d.ts", "protocol"],
+  "scripts": {
+    "build": "node scripts/validate-contracts.js",
+    "test": "node --test test/*.test.js"
+  }
+}
