+++ packages/shared/protocol/index.js
@@ -1,0 +1,5 @@
+const path = require('node:path');
+
+module.exports = {
+  schemaPath: path.join(__dirname, 'schema.proto'),
+};
