+++ pnpm-workspace.yaml
@@ -1,0 +1,17 @@
+packages:
+  - api
+  - client
+  - server
+  - discord-bots
+  - scripts/dbmigrator
+  - packages/*
+
+allowBuilds:
+  '@nestjs/core': true
+  '@parcel/watcher': true
+  bcrypt: true
+  esbuild: true
+  protobufjs: true
+  sharp: true
+  uWebSockets.js: true
+  'uWebSockets.js@https://codeload.github.com/uNetworking/uWebSockets.js/tar.gz/a63031f40f76dc2422e8da736c04217053e9db2b': true
