+++ server/Procfile
@@ -1,1 +1,1 @@
-web: yarn start
+web: pnpm start
