+++ server/package.json
@@ -6,12 +6,13 @@
   "scripts": {
     "start": "node ./src/index.js",
     "dev": "nodemon ./src/index.js --watch ./src",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "node --test",
     "foreverprod": "forever start src/index.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@swordbattle/shared": "workspace:*",
     "dotenv": "^16.1.4",
     "leo-profanity": "^1.7.0",
     "poly-decomp": "^0.3.0",
@@ -21,7 +22,6 @@
     "uuid": "^9.0.0"
   },
   "volta": {
-    "node": "22.23.1",
-    "yarn": "1.22.22"
+    "node": "22.23.1"
   }
 }
