--- server/src/network/protocol/schema.proto
@@ -1,176 +1,0 @@
-syntax = "proto3";
-
-message ClientMessage {
-  bool spectate = 1;
-  bool play = 2;
-  float angle = 3;
-  repeated InputData inputs = 4;
-  MouseData mouse = 5;
-  int32 selectedEvolution = 6;
-  int32 selectedBuff = 7;
-  string chatMessage = 8;
-  bool isPing = 9;
-  string token = 10;
-  string name = 11;
-  string captchaP0 = 12;
-  string captchaP1 = 13;
-  string captchaP2 = 14;
-  string captchaP3 = 15;
-  string captchaP4 = 16;
-  string captchaP5 = 45;
-  string captchaP6 = 46;
-  string captchaP7 = 47;
-  string captchaP8 = 48;
-  string captchaP9 = 49;
-  string captchaP10 = 50;
-  string captchaP11 = 51;
-  bool firstLife = 52;
-  int32 selectedCard = 53;
-  bool openCardSelect = 54;
-  bool rerollCard = 55;
-  int32 selectedUpgrade = 61;
-  bool openUpgradeSelect = 62;
-  bool closeUpgradeSelect = 63;
-  bool skipMajorCard = 56;
-  bool tutorialComplete = 57;
-  int32 tutorialPanel = 58;
-  bool closeCardSelect = 59;
-  // Client-authoritative chest timing-bar result. `optional` so the field has
-  // real presence (absent ≠ 0) — pings/other messages don't clobber it.
-  // 0 = none, 1 = miss, 2 = good, 3 = great, 4 = perfect.
-  optional int32 chestHitZone = 60;
-}
-message InputData {
-  int32 inputType = 1;
-  bool inputDown = 2;
-}
-message MouseData {
-  float angle = 1;
-  int32 force = 2;
-}
-
-message ServerMessage {
-  bool fullSync = 1;
-  int32 selfId = 2;
-  Point spectator = 3;
-  MapData mapData = 4;
-  map<int32, Entity> entities = 5;
-  map<int32, Entity> globalEntities = 6;
-  bool isPong = 7;
-  int32 tps = 8;
-  int32 realPlayersCnt = 9;
-}
-message Shape {
-  int32 type = 1;
-  float x = 2;
-  float y = 3;
-  float angle = 4;
-  int32 radius = 5;
-  repeated Point points = 6;
-}
-message Point {
-  int32 x = 1;
-  int32 y = 2;
-}
-message Buff {
-  int32 level = 1;
-  int32 max = 2;
-  float step = 3;
-}
-message Account {
-  int32 id = 1;
-  string created_at = 2;
-  bool subscription = 3;
-  string subscription_start_date = 4;
-  int32 rank = 5;
-  string clan = 6;
-  bool adSupporter = 7;
-}
-message MapData {
-  int32 x = 1;
-  int32 y = 2;
-  int32 width = 3;
-  int32 height = 4;
-  repeated Entity biomes = 5;
-  repeated Entity staticObjects = 6;
-}
-message Entity {
-  // General
-  int32 id = 1;
-  int32 type = 2;
-  int32 depth = 3;
-  bool removed = 4;
-  Shape shapeData = 5;
-  float healthPercent = 6;
-  float angle = 7;
-  int32 size = 8;
-  bool isAngry = 9;
-
-  // Player
-  string name = 10;
-  string username = 11;
-  Account account = 12;
-  int32 kills = 13;
-  map<int32, int32> flags = 14;
-  int32 biome = 15;
-  int32 level = 16;
-  int32 coins = 17;
-  int32 tokens = 45;
-  int32 nextLevelCoins = 18;
-  int32 previousLevelCoins = 19;
-  int32 upgradePoints = 20;
-  map<int32, Buff> buffs = 21;
-  int32 evolution = 22;
-  map<int32, bool> possibleEvolutions = 23;
-  map<int32, bool> possibleUpgrades = 64;
-  repeated int32 currentUpgrades = 65;
-  bool isAbilityAvailable = 24;
-  bool abilityActive = 25;
-  float abilityDuration = 26;
-  float abilityCooldown = 27;
-  float viewportZoom = 28;
-  string chatMessage = 29;
-  bool isFlying = 30;
-  bool swordFlying = 31;
-  float swordSwingAngle = 32;
-  float swordSwingProgress = 33;
-  float swordSwingDuration = 34;
-  float swordFlyingCooldown = 35;
-  string disconnectReasonMessage = 36;
-  bool pullbackParticles = 37;
-
-  // Chest
-  int32 rarity = 38;
-  // Building
-  int32 width = 39;
-  int32 height = 40;
-  int32 disconnectReasonType = 41;
-  // skin
-  int32 skin = 42;
-
-  bool wideSwing = 43;
-  int32 coinShield = 44; // RESERVED: no longer sent (coin-shield mechanic removed); keep so tag 44 is never reused
-
-  // Swing arc for Cleave card
-  float swordSwingArc = 51;
-  bool swordBoomerangReturning = 52;
-
-  // Card system
-  repeated int32 cardOffers = 46;
-  repeated int32 chosenCards = 47;
-  bool choosingCard = 48;
-  float cardTimer = 49;
-  int32 cardPickNumber = 50;
-  int32 availableUpgrades = 53;
-  int32 rerollsAvailable = 54;
-  repeated int32 skipResults = 55;
-  bool isTutorial = 56;
-  bool swordRaising = 57;
-  bool swordDecreasing = 58;
-  bool offhandRaising = 66;
-  bool offhandDecreasing = 67;
-  int32 activeSelection = 68;
-  int32 abilityCharges = 69;
-
-  // Sword block
-}
